requirements.txt
django==3.2.5
django-environ==0.4.5
djangorestframework ==3.12.4
django-filter==2.4.0